feat: ignore require-time blocks
Currently we report the time to import as an event loop block.

import (require, in cjs) is synchronous, so it makes sense that
it blocks, and reporting it is just reporting the startup time
of the application. Most(tm) node applications will do some
async operation (e.g. a download of a resource, or an http/tcp
listen) before they attempt to do any real work, so waiting until
after the first spin seems reasonable.

This is implemented by defaulting the last_alive to 0 (i.e. we have
never yet been alive), then letting the timer set it as normal. Once
the timer has run once, the behaviour will be the same as before. The
timer won't run until we've cleared our first event loop block; the
one that happens at require time.